+++ R/util-fns_equal.R
@@ -9,8 +9,5 @@ fns_equal <- function(x, y) {
     stop("`x` and `y` must be functions.")
   }
 
-  # Replacement for `base::deparse1`, which is only available in R >= 4.0.0.
-  deparse_1 <- function(z) paste(deparse(z), collapse = " ")
-
-  deparse_1(x) == deparse_1(y)
+  deparse1(x) == deparse1(y)
 }