+++ tests/testthat/test-util-get_file_info.R
@@ -1,6 +1,6 @@
 test_that("get_file_info works", {
-  iris_complete_path <- fs::path_package("extdata", package = "filecacher") |>
-    fs::path("iris_complete.csv") |>
+  iris_complete_path <- system.file("extdata", package = "filecacher") |>
+    file.path("iris_complete.csv") |>
     as.character()
 
   iris_complete_info <- get_file_info(iris_complete_path)
@@ -10,14 +10,16 @@ test_that("get_file_info works", {
   )
   expect_equal(
     iris_complete_info$size,
-    as.character(as.numeric(fs::file_size(iris_complete_path)))
+    as.character(file.info(iris_complete_path)$size)
   )
 
   # Only check `mtime` up to milliseconds:
   #   xxxx-xx-xx xx:xx:xx.xxx
   expect_equal(
-    iris_complete_info$mtime |> substr(1, 23),
-    as.character(fs::file_info(iris_complete_path)$modification_time) |>
+    iris_complete_info$mtime |>
+      substr(1, 23),
+    file.info(iris_complete_path)$mtime |>
+      as.character() |>
       substr(1, 23)
   )
 })