+++ tests/testthat/test-vectorize_reader.R
@@ -1,7 +1,7 @@
 test_that("vectorize_reader works", {
-  extdata_directory <- fs::path_package("extdata", package = "filecacher")
-  iris_complete_path <- fs::path(extdata_directory, "iris_complete.csv")
-  iris_only_paths <- fs::dir_ls(extdata_directory, glob = "*only.csv")
+  extdata_directory <- system.file("extdata", package = "filecacher")
+  iris_complete_path <- file.path(extdata_directory, "iris_complete.csv")
+  iris_only_paths <- fs_dir_ls_(extdata_directory, glob = "*only.csv")
 
   iris_complete <- read.csv(iris_complete_path)
   iris_from_subsets <- vectorize_reader(read.csv)(iris_only_paths)